requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.1
Automat==20.2.0
boto3==1.16.56
botocore==1.19.56
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.0.9
django-cors-headers==3.4.0
django-friendship==1.9.1
django-mysql==3.10.0
django-secure==1.0.2
django-sslserver==0.22
django-storages==1.11.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
msgpack==1.0.1
mysqlclient==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
s3transfer==0.3.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
zope.interface==5.2.0